test: move usage tests
