+++ src/M31.FluentApi.Tests/CodeGeneration/CodeGenerationExecutionTests.cs
@@ -17,208 +17,6 @@ public partial class CodeGenerationTests
 
 
     [Fact, Priority(1)]
-    public void CanExecuteFluentLambdaClass()
-    {
-        var student = TestClasses.Abstract.FluentLambdaClass
-            .CreateStudent
-            .WithName("Alice")
-            .WithAddress(a => a.WithHouseNumber("23").WithStreet("Market Street").InCity("San Francisco"));
-
-        Assert.Equal("Alice", student.Name);
-        Assert.Equal("23", student.Address.HouseNumber);
-        Assert.Equal("Market Street", student.Address.Street);
-        Assert.Equal("San Francisco", student.Address.City);
-    }
-
-    [Fact, Priority(1)]
-    public void CanExecuteFluentLambdaCollectionClass()
-    {
-        var student = TestClasses.Abstract.FluentLambdaCollectionClass
-            .CreateStudent
-            .WithName("Alice")
-            .WithAddresses(
-                a => a.WithHouseNumber("23").WithStreet("Market Street").InCity("San Francisco"),
-                a => a.WithHouseNumber("108").WithStreet("5th Avenue").InCity("New York"));
-
-        Assert.Equal("Alice", student.Name);
-        Assert.Equal(2, student.Addresses.Count);
-        Assert.Equal("23", student.Addresses[0].HouseNumber);
-        Assert.Equal("Market Street", student.Addresses[0].Street);
-        Assert.Equal("San Francisco", student.Addresses[0].City);
-        Assert.Equal("108", student.Addresses[1].HouseNumber);
-        Assert.Equal("5th Avenue", student.Addresses[1].Street);
-        Assert.Equal("New York", student.Addresses[1].City);
-    }
-
-    [Fact, Priority(1)]
-    public void CanExecuteFluentLambdaCompoundClass()
-    {
-        var student = TestClasses.Abstract.FluentLambdaCompoundClass
-            .CreateStudent
-            .WithName("Alice")
-            .WithDetails(
-                a => a.WithHouseNumber("23").WithStreet("Market Street").InCity("San Francisco"),
-                p => p.WithNumber("222-222-2222").WithUsage("CELL"));
-
-        Assert.Equal("Alice", student.Name);
-        Assert.Equal("23", student.Address.HouseNumber);
-        Assert.Equal("Market Street", student.Address.Street);
-        Assert.Equal("San Francisco", student.Address.City);
-        Assert.Equal("222-222-2222", student.Phone.Number);
-        Assert.Equal("CELL", student.Phone.Usage);
-    }
-
-    [Fact, Priority(1)]
-    public void CanExecuteFluentLambdaManyCollectionsClass()
-    {
-        {
-            var student = TestClasses.Abstract.FluentLambdaManyCollectionsClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddressesE(createAddressesE: null);
-
-            Assert.Equal("Alice", student.Name);
-            Assert.Null(student.AddressesE);
-        }
-        {
-            var student = TestClasses.Abstract.FluentLambdaManyCollectionsClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddressesF(createAddressesF: _ => null);
-
-            Assert.Single(student.AddressesF);
-            Assert.Null(student.AddressesF[0]);
-        }
-        {
-            var student = TestClasses.Abstract.FluentLambdaManyCollectionsClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddressesG(createAddressesG: null);
-
-            Assert.Equal("Alice", student.Name);
-            Assert.Null(student.AddressesG);
-        }
-        {
-            var student = TestClasses.Abstract.FluentLambdaManyCollectionsClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddressesG(createAddressesG: _ => null);
-
-            Assert.Equal("Alice", student.Name);
-            Assert.NotNull(student.AddressesG);
-            Assert.Single(student.AddressesG!);
-            Assert.Null(student.AddressesG![0]);
-        }
-    }
-
-    [Fact, Priority(1)]
-    public void CanExecuteFluentLambdaManyPrivateCollectionsClass()
-    {
-        {
-            var student = TestClasses.Abstract.FluentLambdaManyPrivateCollectionsClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddressesE(createAddressesE: null);
-
-            Assert.Equal("Alice", student.Name);
-            Assert.Null(student.AddressesE);
-        }
-        {
-            var student = TestClasses.Abstract.FluentLambdaManyPrivateCollectionsClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddressesF(createAddressesF: _ => null);
-
-            Assert.Single(student.AddressesF);
-            Assert.Null(student.AddressesF[0]);
-        }
-        {
-            var student = TestClasses.Abstract.FluentLambdaManyPrivateCollectionsClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddressesG(createAddressesG: null);
-
-            Assert.Equal("Alice", student.Name);
-            Assert.Null(student.AddressesG);
-        }
-        {
-            var student = TestClasses.Abstract.FluentLambdaManyPrivateCollectionsClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddressesG(createAddressesG: _ => null);
-
-            Assert.Equal("Alice", student.Name);
-            Assert.NotNull(student.AddressesG);
-            Assert.Single(student.AddressesG!);
-            Assert.Null(student.AddressesG![0]);
-        }
-    }
-
-    [Fact, Priority(1)]
-    public void CanExecuteFluentLambdaNullablePropertyClass()
-    {
-        {
-            var student = TestClasses.Abstract.FluentLambdaNullablePropertyClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddress(a => a.WithHouseNumber("23").WithStreet("Market Street").InCity("San Francisco"));
-
-            Assert.Equal("Alice", student.Name);
-            Assert.Equal("23", student.Address!.HouseNumber);
-            Assert.Equal("Market Street", student.Address!.Street);
-            Assert.Equal("San Francisco", student.Address!.City);
-        }
-        {
-            var student = TestClasses.Abstract.FluentLambdaNullablePropertyClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithoutAddress();
-
-            Assert.Equal("Alice", student.Name);
-            Assert.Null(student.Address);
-        }
-        {
-            var student = TestClasses.Abstract.FluentLambdaNullablePropertyClass
-                .CreateStudent
-                .WithName("Alice")
-                .WithAddress(_ => null);
-
-            Assert.Equal("Alice", student.Name);
-            Assert.Null(student.Address);
-        }
-    }
-
-    [Fact, Priority(1)]
-    public void CanExecuteFluentLambdaRecursiveClass()
-    {
-        var student = TestClasses.Abstract.FluentLambdaRecursiveClass
-            .CreateStudent
-            .WithName("Alice")
-            .WithFriend(f => f
-                .WithName("Bob")
-                .WithFriend(f2 => f2
-                    .WithName("Eve")
-                    .WithoutFriend()));
-
-        Assert.Equal("Alice", student.Name);
-        Assert.Equal("Bob", student.Friend!.Name);
-        Assert.Equal("Eve", student.Friend!.Friend!.Name);
-        Assert.Null(student.Friend!.Friend!.Friend);
-    }
-
-    [Fact, Priority(1)]
-    public void CanExecuteFluentLambdaSingleStepClass()
-    {
-        var student = TestClasses.Abstract.FluentLambdaSingleStepClass
-            .CreateStudent
-            .WithAddress(a => a.WithHouseNumber("23").WithStreet("Market Street").InCity("San Francisco"));
-
-        Assert.Equal("23", student.Address.HouseNumber);
-        Assert.Equal("Market Street", student.Address.Street);
-        Assert.Equal("San Francisco", student.Address.City);
-    }
-
-    [Fact, Priority(1)]
     public void CanExecuteFluentMethodClass()
     {
         var student = TestClasses.Abstract.FluentMethodClass
